multi-step form setup.
initially set up an array with states to change visuals and state upon form submissions.

might need to make more states or can make it object instead for the  selections of plans, addons, and then summary we can make it something else. for them to be either stored via the array.