full functioning form as of now!
Within this form I have updated the form-handling process to update and change the follow:

The left side to assure which page you are on.

Input fields to assure they are completed.

Add the buttons to handle the change to go to the next process/page.

Within those being complete time to design based off the wireframes.