Button for the form including step validation, validation of the length of form
Within this I assured the state and steps of the form were in order using useContext and the stepContext logic created yesterday to go from first to the last step.

using the step context and form context logic. Validation of the case switch via a function and dispatch the steps if the types are accordingly.