Addons and FormHeader Components
Addons were completed by const variables checking isAdded for the state of addons and then checking if the state isYearly or monthly to validate pricing.

Within inputs I will handled change and checked for isAdded to depend the state of addons id and name.

form header component was to show props title and text to display them upon the header.