FormAddon, FormPersonal, FormPlan, Plan, and Switch Components
FormAddons Component was created with mapping addons.

FormPersonal Component was created to validate inputs and useState for the errors. also using useEffect to handleChange, Blur and set validations.

FormPlan was created to handle change via event and map out the plan Selection!

Plan was created to map our plans and also check if state is yearly or monthly to display such items.

Switch component created for users to switch between yearly or monthly billing in which it will change display on change via props.